checkpoint: moves "/Encoded/exports.ts" to "/Encoded/Compatibility/exports.ts"
