+++ src/library/Sequence/Byte/Encoded/index.ts
@@ -1,1 +1,1 @@
-export * as Sequence_Byte_Encoded from './exports.ts';
+export * as Sequence_Byte_Encoded from './Compatibility/exports.ts';
